facor out parsing and sorting code to a helper
